+++ apps/server/src/config.ts
@@ -24,6 +24,9 @@ export const configSchema = t.Intersect([
 			REDIS_PORT: t.String({ default: "6379" }),
 			BACKEND_PORT: t.String({ default: "3000" }),
 			REDIS_PASSWORD: t.String(),
+			REDIS_TLS_CA: t.String({ default: "/tls/ca.crt" }),
+			REDIS_TLS_CERT: t.String({ default: "/tls/client.crt" }),
+			REDIS_TLS_KEY: t.String({ default: "/tls/client.key" }),
 		}),
 	),
 ]);