+++ apps/server/src/libs/redis.ts
@@ -1,8 +1,17 @@
+import { readFileSync } from "node:fs";
 import { RedisClient } from "bun";
 import { config } from "@/config";
 
 export const redisClient = new RedisClient(
-	`redis://default:${config.REDIS_PASSWORD}@${config.REDIS_HOST}:${config.REDIS_PORT}`,
+	`rediss://:${config.REDIS_PASSWORD}@${config.REDIS_HOST}:${config.REDIS_PORT}`,
+	{
+		tls: {
+			ca: readFileSync(config.REDIS_TLS_CA!),
+			cert: readFileSync(config.REDIS_TLS_CERT!),
+			key: readFileSync(config.REDIS_TLS_KEY!),
+			rejectUnauthorized: true,
+		},
+	},
 );
 
 redisClient.onconnect = async () => {